refactor(edge): isolate metrics endpoint rendering and request handling
